fix(tui): 移除不必要的 u16 -> u16 类型转换
clippy unnecessary_cast 警告：scroll_offset_y 已是 u16

+++ peri-tui/src/kit/message_area/mod.rs
@@ -500,7 +500,7 @@ pub fn MessageArea(props: &MessageAreaProps, mut hooks: Hooks) -> impl Into<AnyE
             Text(
                 text: Paragraph::new(RatText::from(viewport_lines))
                     .wrap(Wrap { trim: false }),
-                scroll: Position::new(scroll_offset_y as u16, 0),
+                scroll: Position::new(scroll_offset_y, 0),
             )
         }
     )
